aula22

+++ mundo 3/aula 20/atividade96.py
@@ -1,4 +1,4 @@
-largura = int(input("digite a largura: ")) 
+largura =float(input("digite a largura: ")) 
 comprimento = int(input("digite o comprimento: "))
 def area(largura,comprimento):
     area =largura * comprimento
